Jcw87 to Sam
The project files for VS 2005 are still outdated/messed up. I have a patch that will bring the VS 2005 project files up to date with the VS 2008 project files, add the x64 platform to the VS 2005 project files, fix a few issues with x64 in VS 2008, and a few other misc changes that don't really change the output at all.