Optionally don't clear the fbcon framebuffer on shutdown.
The new behavior is triggered by setting the environment variable
SDL_FBCON_DONT_CLEAR.

This is useful in certain circumstances (specifically: a game launcher on an
embedded device can leave a "now loading!" screen on the framebuffer while
another process is starting up).

Fixes Bugzilla #1251.

Thanks to Paul Cercueil for the patch!